requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
youtube-dl==2021.12.17
yt-dlp==2023.7.6
python-dotenv==1.0.0
pillow==10.0.0
requests==2.31.0
aiohttp==3.8.5
aiofiles==23.1.0
pydub==0.25.1 